Boost - update external to 1.74
Not sure why the previous update attempt didn't work - trying again.

+++ app/gui/qt/external/scsynth-boost-1.74.0/boost/test/utils/runtime/cla/argv_traverser.hpp
@@ -17,6 +17,7 @@
 
 // Boost.Test Runtime parameters
 #include <boost/test/utils/runtime/fwd.hpp>
+#include <cstring>
 
 #include <boost/test/detail/suppress_warnings.hpp>
 
@@ -47,7 +48,7 @@ public:
     /// Returns new argc
     int         remainder()
     {
-        return m_argc;
+        return static_cast<int>(m_argc);
     }
 
     /// Returns true, if we reached end on input
